renaming mapping file
