+++ src/bridge/pipelines/mapping/__init__.py
@@ -2,7 +2,7 @@
 Classes for mapping metadata records from one schema to another.
 """
 
-from .main import MapBioTools2GitHub
+from .bt2gh_map import MapBioTools2GitHub
 
 __all__ = [
     "MapBioTools2GitHub",
