+++ src/bridge/pipelines/mapping/bt2gh_map.py
